requirements.txt
datasets
langdetect
langid
fasttext-wheel
pandas